cloud_functions/youtube_listener/requirements.txt
functions-framework>=3.0.0
google-api-python-client>=2.0.0
google-cloud-bigquery>=3.0.0
python-dateutil>=2.8.0
google-cloud-storage>=2.0.0
vaderSentiment
nltk
pyahocorasick
